// frontend/src/api/models/mail-account.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MailAccount {
  createdAt: string;

  /**
   * Unique identifier
   */
  id: string;

  /**
   * IMAP server hostname
   */
  imapHost: string;

  /**
   * IMAP server port
   */
  imapPort: number;

  /**
   * Human-readable account name
   */
  name: string;

  /**
   * Transport security used for the IMAP connection
   */
  security: 'ssl' | 'starttls' | 'none';
  updatedAt: string;

  /**
   * IMAP login username
   */
  username: string;
}